Fix Dashboard counts in localsite for new user
CounterField's _reinit method was returning back control if the pk
of the model wasnt setup yet. This caused any of the "initializers"
to not fire for the very first time. Modified the logic so that
control returns back if the model's pk is None and there are no
initializers defined.

For the case of LocalSiteProfiles table, there are initializers
defined for all CounterFields, so they would be initialized the first
time the object is created.

Fixes bug2873